Замена int на long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
 
     public static void main(String[] args) {
-        SalesManager version = new SalesManager(new int[]{100, 200, 300});
+        SalesManager version = new SalesManager(new long[]{100, 200, 300});
         System.out.println(version.max());
     }
 }
